[giac] update to giac 1.1.4, many changes, keep your fingers crossed...
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@39339 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/giac/first.cc
@@ -90,7 +90,7 @@ init_gmp_memory::~init_gmp_memory() { }
 
 #include <iostream>
 
-unsigned long giac_allocated = 0;
+size_t giac_allocated = 0;
 void* operator new(std::size_t size)
 {
   std::cerr << giac_allocated << " + " << size << std::endl;
